requirements.txt
google-adk>=0.5.0
xgboost>=2.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy<2.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
